Add support for loading platform dependent tale scenes
When loading a tale we now check if there is a platform dependent scene
class and load that instead of the generic one.

The class must be in a package adjacent to the generic class, where the
package name is the platform name and the class name is the generic
class name prefixed by the platform name.