Implement creating an Injector in the engine setup
The Injector is created from the IInjectConfig in the setup and passed
to the engine context.
For each subproject there is now a main Guice module and a default
IInjectConfig that installs all necessary modules by extending and
overriding as needed.

Android and Desktop launchers now take an IInjectConfig at creation.