Add tokTales-test project
This project and it's subprojects are the base for testing. They contain
all test-library dependencies and various helper classes. Other projects
will use them as their test dependencies.

They also contain classes that are meant to ensure dependency integrity,
as in having correct access to the types provided by dependencies.