Add ParameterInjector files including tests
The parameter injector can be used to inject parameters (specific
instances) into objects of unknown type, after their creation.

The objects are scanned for methods marked with certain annotations
which are then invoked with matching parameters. A generic annotation
InjectParameters is provided, that can also be customized further with a
key indicating which injector should invoke which methods.