Implement support for parameter injection on Android
Android versions before API Level 24 do not support the Repeatable
annotation or any newer methods for getting annotations.

Added an annotation CompatRepeatable that can be used in addition to
Repeatable to ensure compatibility.
Also implemented a manual way of getting repeatable annotations that
does not use any of the newer methods.