Refactor ParameterInjector to check subclass methods first
Previously, superclass methods were discovered first and immediately
invoked. Now the injector will first scan for methods from the first
class through to the top superclass, and ignore superclass methods it
has already checked in subclasses.
As a result an inherited method not annotated with an injection
annotation will not be invoked, even if the superclass is annotated.
This behavior is intended and matches the way Guice works.

Also fixes the fact that inherited methods would be invoke twice, once
for the class and once for the superclass.